requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
discord.py==2.2.2
fonttools==4.39.3
frozenlist==1.3.3
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.2
packaging==23.0
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
urllib3==1.26.15
yarl==1.8.2
